# codebase/src/api_service/out/openapi_server/models/login_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class LoginResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, token=None):  # noqa: E501
        """LoginResponse - a model defined in OpenAPI

        :param token: The token of this LoginResponse.  # noqa: E501
        :type token: str
        """
        self.openapi_types = {
            'token': str
        }

        self.attribute_map = {
            'token': 'token'
        }

        self._token = token

    @classmethod
    def from_dict(cls, dikt) -> 'LoginResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The LoginResponse of this LoginResponse.  # noqa: E501
        :rtype: LoginResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def token(self) -> str:
        """Gets the token of this LoginResponse.

        JWT token returned to the client  # noqa: E501

        :return: The token of this LoginResponse.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token: str):
        """Sets the token of this LoginResponse.

        JWT token returned to the client  # noqa: E501

        :param token: The token of this LoginResponse.
        :type token: str
        """

        self._token = token