# codebase/src/api_service/out/openapi_server/models/product.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Product(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, name=None, description=None, category=None, price=None, slogan=None, stock=None, created_at=None):  # noqa: E501
        """Product - a model defined in OpenAPI

        :param id: The id of this Product.  # noqa: E501
        :type id: int
        :param name: The name of this Product.  # noqa: E501
        :type name: str
        :param description: The description of this Product.  # noqa: E501
        :type description: str
        :param category: The category of this Product.  # noqa: E501
        :type category: str
        :param price: The price of this Product.  # noqa: E501
        :type price: float
        :param slogan: The slogan of this Product.  # noqa: E501
        :type slogan: str
        :param stock: The stock of this Product.  # noqa: E501
        :type stock: int
        :param created_at: The created_at of this Product.  # noqa: E501
        :type created_at: datetime
        """
        self.openapi_types = {
            'id': int,
            'name': str,
            'description': str,
            'category': str,
            'price': float,
            'slogan': str,
            'stock': int,
            'created_at': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'category': 'category',
            'price': 'price',
            'slogan': 'slogan',
            'stock': 'stock',
            'created_at': 'created_at'
        }

        self._id = id
        self._name = name
        self._description = description
        self._category = category
        self._price = price
        self._slogan = slogan
        self._stock = stock
        self._created_at = created_at

    @classmethod
    def from_dict(cls, dikt) -> 'Product':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Product of this Product.  # noqa: E501
        :rtype: Product
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Product.


        :return: The id of this Product.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Product.


        :param id: The id of this Product.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Product.


        :return: The name of this Product.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Product.


        :param name: The name of this Product.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Product.


        :return: The description of this Product.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Product.


        :param description: The description of this Product.
        :type description: str
        """

        self._description = description

    @property
    def category(self) -> str:
        """Gets the category of this Product.


        :return: The category of this Product.
        :rtype: str
        """
        return self._category

    @category.setter
    def category(self, category: str):
        """Sets the category of this Product.


        :param category: The category of this Product.
        :type category: str
        """

        self._category = category

    @property
    def price(self) -> float:
        """Gets the price of this Product.


        :return: The price of this Product.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this Product.


        :param price: The price of this Product.
        :type price: float
        """

        self._price = price

    @property
    def slogan(self) -> str:
        """Gets the slogan of this Product.


        :return: The slogan of this Product.
        :rtype: str
        """
        return self._slogan

    @slogan.setter
    def slogan(self, slogan: str):
        """Sets the slogan of this Product.


        :param slogan: The slogan of this Product.
        :type slogan: str
        """

        self._slogan = slogan

    @property
    def stock(self) -> int:
        """Gets the stock of this Product.


        :return: The stock of this Product.
        :rtype: int
        """
        return self._stock

    @stock.setter
    def stock(self, stock: int):
        """Sets the stock of this Product.


        :param stock: The stock of this Product.
        :type stock: int
        """

        self._stock = stock

    @property
    def created_at(self) -> datetime:
        """Gets the created_at of this Product.


        :return: The created_at of this Product.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at: datetime):
        """Sets the created_at of this Product.


        :param created_at: The created_at of this Product.
        :type created_at: datetime
        """

        self._created_at = created_at